Test adding and removing 10,000 files at a time.
It turns out that recent versions of Perl 6 handle buffering
the output of a command automatically, so the hang when we're
writing to the 'xargs | git' pipe and git is reporting stuff
back no longer happens.  So... do not fix it, but depend on it,
just test it.  Yeah, yeah, I know.